fix: lined-updated properties panel with switching tabs
